Arithmetic
I started implementing the arithmetic instructions. For now I will
not implement BCD mode. I will leave the door for BCD open but I will
not enter it now. I am experiencing a strnage bug on the
macros that do the arithmetic instructions. Specifically with the SUB
one. I don't entirely know what the issue, I just know that it is not
working.