CPU Switch
I changed what CPU I am going to use in the real life build, so I must
change the emulator accordingly. The CPU I am going to use is the 65C02.
Flags are also now being set as they should be. I just need to do the
rest of the logical instructions and the additional instructions that
the 65C02 implements.