Cycle Measurement
I would like to be able to adjust the frequency of the emulator: i.e. I
want it to run at 10 MHz, how many cycles will I need a second to
complete 10 MHz, and from that be able to produce the number of times
I need to call tick_65C02 to accomplish it.