Sprites
I have implemented sprites, they basically use slightly modified
terminal code. I am considering moving the video register into the zero-
page as it is fairly small and video is going to be accessed quite a
bit.
For now, I am just drawing the first 250 bytes of the kernel on the top
left of the screen.