Commit
I have fixed some tiny logic issues within instructions suffixed with
_ABS_X, I have also double checked my table of cycle counts to ensure
that they are correct. I did not want to do this manually so I wrote a
tool to do this. I have provided the instructions in a CSV file as found
from https://www.zophar.net/fileuploads/2/10533qqcap/6502ref.html.
The next step is again with cycle counts, and this is just to ensure
that if an instruction adds additional cycles
dependant on circumstance, it is accounted for.