C Kernal
I got the C kernal running. Warped my mind a lot but I am glad I did it.
The solution is quite simple, and I realize now that I just had to load
the library.
To do that I needed to READ THE MANUAL. Which I am not too good at.
Anyway, it all works now. Unfortunately, no optimization and we are
already taking up 1/4 kilobyte of the 8 kilobytes we
have to work with, which is not ideal. I am going to try to find
a solution to optimize the code and still have it work.