cc65 madness
I was trying to write a mother's day program in the kernal.
IT JUST CAN'T DO IT.
cc65 just doesn't seem to be able to do compile the code so that it
works.
I don't know if this is my fault, nor do I know if this is the
fault of the compiler.
All I know is that it is just not working, and I do not know how to fix
it.