Program Loading
I added a function which would load a program. This function
basically wraps the disk loading function. And checks that the
loaded data starts with the word $ECEC.
I also added a new feature to the video chip which is the option to
disable drawing the foreground.
I noticed some artifacting in the video, which I want to fix.