Logo loaded
I have made some code to draw a logo. There is a python script to
generate the BMP image from a file, and well that's about it.
The assembly code just draws what is given to it where it is told to
draw it.
I also changed it so that the X and Y coordinates of sprites are encoded
in X: low byte, Y: high byte order, instead of an index. There was also
some corrections made as to how the address of the sprite is calculated
and to draw it the right way around.