Terminal changes
Removed the C files, this will be done in assembly.
Handling of the terminal has been moved to the main program loop,
IRQs now only handle reporting what key was pressed and stating
if there is something new to handle such as a backspace, enter or
regular character.
The IRQ handler now also skips over the IRQ terminal handling part
if a program (besides the kernal) is recognized to be executing.
A bug with checking the RTC has been fixed. The logo screen would
forever display if a minute boundary was crossed during the wait.
This is now fixed with using the amount of unique seconds that have
passed.