Video changes
I have separated the video buffers. This is to allow for hardware
scrolling. The ability for each character on the screen to have its own
foreground or background is no longer the case. There is also now an
extra bit in the video control register status field to enable and
disable hardware scrolling.
Due to changes made to the rendering function, some faults do arise when
closing the program. These faults are noted down in the "destroy_video"
function.