Working on larger files
Been working on the logic to make the file system be able to write
larger files. So far, it is not working as it writes over previous data.
I have no idea why it is doing this, it should simply not be possible.
The code to write more than 1 KB looks fine.
There is also a new attribute bit "Exists" that allows for the driver to
differentiate between non-written and written
entries.