Fixed and deleting files
The issue ended up being me using a uint16_t instead of a long in the
write_to_fs function. I also had to add some if statements to detect if
we are writing the last block of a file, and if so to not update some
sector related stuff.
I also made a function to mark files as deleted.