Kernal and file system
Started working on the assembly file system driver.
Removed some files to make the repository a little bit
cleaner.
It seems that when a new directory needs to be allocated in the
file system, the initial directory is zeroed.
The binary files (previously .fs files) contain the file systems.
There are two:
data.bin : Testing the prototype driver (written in C).
test.bin : Testing the kernal file system driver.