Major errors and directory listing
Fixed some major errors within the emulator that I managed to gloss
over.
Added support for the kernal to print more than one file when
the DIR command is invoked. The initialization of the vfs in the kernal
was also fixed, so instead of loading 1024 sectors, it now loads the
amount of 2.