Lots of fixes and disk delays
Fixed some things, most of these files just things the linter was
complaining about. A scheduler was made, this is because I did not want
to have two threads to tick the control register and CPU. The scheduler
was implemented for emulating disk delays that may occur in real life.
I had written some code that would account for this delay in assembly,
but the problem was that the interrupt was being called in negative time
before the CPU could even get to WAI statement. Slowing the disk
operation down to take at least 250 milliseconds is annoying, but I feel
adds a layer of realism. It would also be simpler as real life delay is
accounted for in the emulator, thus the kernal versions that are in a
physical ROM or an emulator do not after to differ.