Finding and reading files
Created a memcpy routine and I am currently implementing fs_load_file.
When trying to load the data of any other file that is not TEST.TXT,
there is an error. The error is that the first file entry of the initial
directory is printed instead of the real file data. Upon further
investigation, this is because fs_find_file cannot find any other file
than TEST.TXT.