DIR and terminal oversight
Fixed the DIR command, it is now compatible with hardware scrolling.
Also discovered that the root of the problem for the DIR command not
being executed after a LOAD command is because the terminal buffer is
not cleared. This is now fixed and after every enter the terminal
buffer is cleared. Also documentation has been added. Documentation for
the enter loop will be added in an upcoming commit.