Sharing memory
Currently trying to make it so that the emulator can share its memory
with other programs. This way it would be easier to debug code
and do various other things, such as modifying the contents of registers
, changing memory around, doing memdumps, so on. So far, it is not
going too well. Receiving a segmentation fault when trying to access
the memory that the program has supposedly allocated as shared. Going to
rework the locking mechanism as well.