One little ampersand
In inst_stack.c, when pushing the flags register, least significant
byte of the variable containing the pointer to the flags register in
shared memory was being pushed onto the stack. Everything seems to be
working now. Along with this a little bug is present within the kernal
that allows the user to roll over the X coordinate, but this only works
if nothing has been typed into the terminal.