Sharing memory works
Wrote a little test program to act as the monitor. It printed the kernal
ROM and I currently modified it to print the first 16 bytes of RAM and
it did both correctly. The spinlock remains untested. However the
lock was unique to each program, it is now a pointer to a byte at the
end of the allocated RAM. Hopefully one byte is enough for it and so
far there haven't been issues with the emulator acquiring and releasing
the lock.