Disassembling and more parameters
Added some more parameters to the monitor program. Currently working
testing the disassembling, so far it is not going well. The hash function
doesn't work, but it is able to parse in a labels file. Opening a file
after parsing a labels file simply causes a segmentation fault for some
reason as well.