Few fixes
I noticed that the disassembler was sometimes messing up. This was
due to it being out of synchronization with the CPU. This was fixed  by
just keeping track of the address of the start of the current
instruction. Also made it so that the target instructions per second is
now apart of memory. Going to be working more on the monitoring
interface but so far it is looking good.