Margining system
Implemented some macros to define where the code / label columns
should be drawn as well as implementing a cut off system for strings
which do not fit inbetween the margins. Next I will work on some
indicators. This will highlight the current instruction in a white color
, show the flags, and basically color coding things to make them nicer
to look at.