Issues and fixes
Numerous issues arose when I tried to implement a locking mechanism
in the main loop of the debugger. The reasoning behind the locking
mechanism was so that while the debugger did not have authority over
the memory it would simply pause. However, from this logic rose numerous
deadlocks. This was fixed by the use of a variable for each client to
have which states whether or not it owns the lock. Acquiring the lock
now also pauses instead of just being in a loop. There was also no
difference between the client and host's mechanism to destroy the shared
memory, thus they were unified. The discovery of this was also aided by
not being able to start the debugger after it had been first closed due
to the error of "Failed to initialize memory". Strangely, this was fixed
by not closing the memory on the client's end.