More debugger and initializer
Added another window to the debugger, this one shows some
emulator properties. Currently it only shows the IPS. The IPS has also
been changed to a uint64_t. There are also some hard coded keys that
are capable of changing the speed of the emulator. Also worked on the
code that generates a new directory initializer.