File system error messages
Made it so that initialize_fs checks if it supports the file system
before returning. If it does not support the file system on the disk
it will print one of two error messages.
"File system version not supported!" if an unsupported version of SFS
is being used or "Unsupported file system!" if it could not recognize
the file system as SFS.