Fixed numerous errors
Yesterday I wrote a lot of code, most of it was fine; however, there
were tiny mistakes that rendered the debugger useless. Those have been
fixed. With breakpoints there is a little issue where the emulator slips
off the instruction we are actually supposed to break on, this can be
fixed through the use of some additional checks.