Command prompt
Made it draw out to the user, the previous hashing function didn't
work too well, replaced it with a new one (I have no idea what I am
doing in terms of hashing). The flags that are used by the disassembler
are now a global variable, no more pointer tomfoolery. Need to fix the
slipping issue, where the breakpoint slips off from the instruction it
is supposed to be on to the next instruction.