Backspaces, allocation hygiene
Backspaces are for whatever reason converted into BEL characters. Simple
patch for now is that BEL and BS characters both act as backspace. The
handler for the break command was not very hygienic when it came to
retreiving the symbol that needs to be broken. Also made it so that in
the hexdump, the spaces would remain as the background color.