Weird assembly
For some reason, when trying to create a new directory descriptor,
the file initializer's version number is set to 0. I do not understand
why this would be happening. There should be no code setting its value
to 0. The host file system manager for SFS is also extremely broken when
trying to create another directroy.